feat: add `IS_IPB` packet
You can use the IS_IPB packet to set up to 120 IPv4 addresses that are not allowed to join a host.

The IP addresses assigned to the `BanIPs` property must be in a string format `X.X.X.X` where `X` is a number from 0 to 255. If some of the IP addresses are invalid, a `RangeError` will be thrown.

fix #47